perf: Create DragAndDropSortableList component
