+++ packages/frontend/src/views/admin/views/core/dashboard/types.ts
@@ -1,21 +1,13 @@
 import { UniqueIdentifier } from '@dnd-kit/core';
-import { MutableRefObject } from 'react';
 
-export interface TreeItem {
-  children: TreeItem[];
+export interface TreeItem<T extends TreeItem<T>> {
+  children: T[];
   collapsed?: boolean;
   id: UniqueIdentifier;
 }
 
-export type TreeItems = TreeItem[];
-
-export interface FlattenedItem extends TreeItem {
+export interface FlattenedItem<T extends TreeItem<T>> extends TreeItem<T> {
   depth: number;
   index: number;
   parentId: null | UniqueIdentifier;
 }
-
-export type SensorContext = MutableRefObject<{
-  items: FlattenedItem[];
-  offset: number;
-}>;
